* gmp-impl.h (mpn_zero_p): Remove, no longer needed.
Well, mpn_zero_p was been a bit of a debacle.  It started life as an
idea to share code, but I now think it really ought to be always
inlined and that means coding it explicitly for the benefit of
compilers without "inline".  Those places using it just for
convenience, including the couple of ASSERTs, can easily enough make
other arrangements (and now have done).