Remove usqr_ppm, decide against this.
The only apparent current use of a square is mpn/generic/sqr_basecase.c.
Ultrasparc doesn't reach that, because it has its own sqr_diagonal.
m68000 does reach it, but saving a couple of insns for that chip isn't
worth bothering about.  An assembler sqr_basecase would be the way to
go if it really mattered.